requirements.txt
Django==5.0.6
django-crispy-forms==2.1
matplotlib==3.9.0
reportlab==4.2.0
pandas==2.2.2
numpy==1.26.4
openai==1.31.1
python-dotenv==1.0.1